Add custom username validating rule
